+++ core/execution/future/market_snapshot.go
@@ -84,8 +84,8 @@ func NewMarketFromSnapshot(
 
 	as := monitor.NewAuctionStateFromSnapshot(mkt, em.AuctionState)
 
-	if vgcontext.InProgressUpgradeFrom(ctx, "v0.73.10") {
-		// protocol upgrade from v0.73.9, lets populate the new liquidity-fee-settings with a default marginal-cost method
+	if vgcontext.InProgressUpgradeFrom(ctx, "v0.73.12") {
+		// protocol upgrade from v0.73.12, lets populate the new liquidity-fee-settings with a default marginal-cost method
 		log.Info("migrating liquidity fee settings for existing market", logging.String("mid", mkt.ID))
 		mkt.Fees.LiquidityFeeSettings = &types.LiquidityFeeSettings{
 			Method: types.LiquidityFeeMethodMarginalCost,
@@ -202,7 +202,7 @@ func NewMarketFromSnapshot(
 	now := timeService.GetTimeNow()
 	marketType := mkt.MarketType()
 
-	markPriceCalculator := NewCompositePriceCalculatorFromSnapshot(em.CurrentMarkPrice, em.MarkPriceCalculator)
+	markPriceCalculator := NewCompositePriceCalculatorFromSnapshot(ctx, em.CurrentMarkPrice, timeService, oracleEngine, em.MarkPriceCalculator)
 	market := &Market{
 		log:                           log,
 		mkt:                           mkt,
@@ -249,8 +249,11 @@ func NewMarketFromSnapshot(
 		markPriceCalculator:           markPriceCalculator,
 	}
 
+	markPriceCalculator.setOraclePriceScalingFunc(market.scaleOracleData)
+
 	if em.IndexPriceCalculator != nil {
-		market.indexPriceCalculator = NewCompositePriceCalculatorFromSnapshot(nil, em.IndexPriceCalculator)
+		market.indexPriceCalculator = NewCompositePriceCalculatorFromSnapshot(ctx, nil, timeService, oracleEngine, em.IndexPriceCalculator)
+		market.indexPriceCalculator.setOraclePriceScalingFunc(market.scaleOracleData)
 	}
 
 	for _, p := range em.Parties {