+++ datanode/sqlstore/migrations/0069_teams_allow_list.sql
@@ -1,8 +1,0 @@
--- +goose Up
-ALTER TABLE teams
-  ADD COLUMN IF NOT EXISTS allow_list JSONB;
-
--- +goose Down
-
-ALTER TABLE teams
-  DROP COLUMN IF EXISTS allow_list;